+++ tests/testthat/test-kpb.R
@@ -19,7 +19,7 @@ test_that("regular progress works", {
       pb <- progress_estimated(5)
       withr::with_output_sink("test_pb.txt",
                 run_pb())
-      expect_match(paste(suppressMessages(readLines("test_pb.txt", warn = FALSE)), collapse = ""), "|========================================================================================================|100% ~0 s remaining")
+      expect_match(scan("test_pb.txt", what = character(), quiet = TRUE), "|========================================================================================================|100% ~0 s remaining")
       file.remove("test_pb.txt")
     }